fix: aggregated data depends on filters
